BinarFood Challenge 6 - Final commit (but not really)

+++ Challange6/binarfood-service/src/main/java/com/kinamulen/binarfood/controller/MerchantController.java
@@ -7,6 +7,7 @@ import com.kinamulen.binarfood.dto.merchant.response.GetMerchantWebResponse;
 import com.kinamulen.binarfood.dto.merchant.response.MerchantWebResponse;
 import com.kinamulen.binarfood.service.InvoiceService;
 import com.kinamulen.binarfood.service.MerchantService;
+import com.kinamulen.binarfood.util.SecurityUtil;
 import lombok.extern.slf4j.Slf4j;
 import net.sf.jasperreports.engine.JRException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,16 +33,20 @@ public class MerchantController {
     private MerchantService merchantService;
     @Autowired
     private InvoiceService invoiceService;
+    @Autowired
+    private SecurityUtil securityUtil;
 
-    @PostMapping
+    @PostMapping("/register/_public")
     public ResponseEntity<MerchantWebResponse> create(@RequestBody CreateMerchantWebRequest request){
         log.info("Starting merchant register, merchant name: {}", request.getMerchantName());
         MerchantWebResponse response = merchantService.create(request);
         return ResponseEntity.ok(response);
     }
 
+
+
     //Get all merchants (including closed merchants)
-    @GetMapping
+    @GetMapping("/_public")
     public ResponseEntity<List<MerchantWebResponse>> getMerchants(@RequestHeader(value = "page", required = false, defaultValue="0") Integer page,
                                                                   @RequestHeader(value = "size", required = false, defaultValue="10") Integer size,
                                                                   @RequestHeader(value = "sortBy", required = false, defaultValue="createdAt") String sortBy,
@@ -51,14 +56,19 @@ public class MerchantController {
     }
 
     //Get open merchants only
-    @GetMapping("/open")
+    @GetMapping("/open/_public")
     public ResponseEntity<List<MerchantWebResponse>> getOpenMerchants() {
         List<MerchantWebResponse> responses = merchantService.getOpenMerchants();
         return ResponseEntity.ok(responses);
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<GetMerchantWebResponse> getMerchant(@PathVariable UUID id) {
+    @GetMapping("/{id}/_merchant-secured")
+    public ResponseEntity<GetMerchantWebResponse> getMerchant(
+            @PathVariable UUID id,
+            @RequestHeader(value = "userId") String idFromToken) {
+        if (Boolean.FALSE.equals(securityUtil.authorizeId(id, idFromToken))){
+            return ResponseEntity.badRequest().build();
+        }
         GetMerchantWebResponse response = merchantService.getMerchant(id);
         if (Objects.nonNull(response)) {
             return ResponseEntity.ok(response);
@@ -66,10 +76,15 @@ public class MerchantController {
         return ResponseEntity.notFound().build();
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("/{id}/_update/_merchant-secured")
     public ResponseEntity<MerchantWebResponse> update(
-            @PathVariable UUID id, @RequestBody UpdateMerchantWebRequest updateMerchantWebRequest) {
+            @PathVariable UUID id,
+            @RequestBody UpdateMerchantWebRequest updateMerchantWebRequest,
+            @RequestHeader(value = "userId") String idFromToken) {
         log.info("Starting merchant update, merchant name: {}", updateMerchantWebRequest.getMerchantName());
+        if (Boolean.FALSE.equals(securityUtil.authorizeId(id, idFromToken))){
+            return ResponseEntity.badRequest().build();
+        }
         MerchantWebResponse response = merchantService.updateMerchant(id, updateMerchantWebRequest);
         if (Objects.nonNull(response)) {
             return ResponseEntity.ok(response);
@@ -77,9 +92,13 @@ public class MerchantController {
         return ResponseEntity.notFound().build();
     }
 
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Boolean> delete(@PathVariable UUID id) {
-        //to do: pake field isDeleted
+    @DeleteMapping("/{id}/_delete/_merchant-secured")
+    public ResponseEntity<Boolean> delete(
+            @PathVariable UUID id,
+            @RequestHeader(value = "userId") String idFromToken) {
+        if (Boolean.FALSE.equals(securityUtil.authorizeId(id, idFromToken))){
+            return ResponseEntity.badRequest().build();
+        }
         boolean response = merchantService.deleteMerchant(id);
         if (response) {
             return ResponseEntity.ok(true);
@@ -87,11 +106,16 @@ public class MerchantController {
         return ResponseEntity.notFound().build();
     }
 
-    @PostMapping("/{id}/report")
-    public ResponseEntity<Resource> getMerchantReport(@PathVariable UUID id,
-                                                      @RequestBody MerchantReportWebRequest request)
+    @PostMapping("/{id}/report/_merchant-secured")
+    public ResponseEntity<Resource> getMerchantReport(
+            @PathVariable UUID id,
+            @RequestBody MerchantReportWebRequest request,
+            @RequestHeader(value = "userId") String idFromToken)
             throws JRException, IOException {
 
+        if (Boolean.FALSE.equals(securityUtil.authorizeId(id, idFromToken))){
+            return ResponseEntity.badRequest().build();
+        }
         byte[] reportContent = invoiceService.genereteReport(id, request.getStartDate(), request.getEndDate());
         ByteArrayResource resource = new ByteArrayResource(reportContent);
         return ResponseEntity.ok()
